lastgenre: Add another test checking fallback
Check if for whatever reason (it usually should not happen) empty
strings in a list are received. In that case _get_genre should move on
to the next stage, which is falling back to the original genre.

Also note that even though whitelist is in place, this will happen,
resulting in an original genre that might not be in the whitelist. This
is expected!